web: cleanup
