+++ web/src/flow.ts
@@ -1,3 +1,3 @@
 import "construct-style-sheets-polyfill";
 
-import "./pages/generic/FlowExecutor";
+import "./flows/FlowExecutor";
