+++ web/src/flows/FlowExecutor.ts
@@ -1,34 +1,34 @@
 import { gettext } from "django";
 import { LitElement, html, customElement, property, TemplateResult, CSSResult, css } from "lit-element";
 import { unsafeHTML } from "lit-html/directives/unsafe-html";
-import "../../elements/stages/authenticator_static/AuthenticatorStaticStage";
-import "../../elements/stages/authenticator_totp/AuthenticatorTOTPStage";
-import "../../elements/stages/authenticator_validate/AuthenticatorValidateStage";
-import "../../elements/stages/authenticator_webauthn/WebAuthnAuthenticatorRegisterStage";
-import "../../elements/stages/autosubmit/AutosubmitStage";
-import "../../elements/stages/captcha/CaptchaStage";
-import "../../elements/stages/consent/ConsentStage";
-import "../../elements/stages/email/EmailStage";
-import "../../elements/stages/identification/IdentificationStage";
-import "../../elements/stages/password/PasswordStage";
-import "../../elements/stages/prompt/PromptStage";
-import { ShellChallenge, RedirectChallenge } from "../../api/Flows";
-import { IdentificationChallenge } from "../../elements/stages/identification/IdentificationStage";
-import { PasswordChallenge } from "../../elements/stages/password/PasswordStage";
-import { ConsentChallenge } from "../../elements/stages/consent/ConsentStage";
-import { EmailChallenge } from "../../elements/stages/email/EmailStage";
-import { AutosubmitChallenge } from "../../elements/stages/autosubmit/AutosubmitStage";
-import { PromptChallenge } from "../../elements/stages/prompt/PromptStage";
-import { AuthenticatorTOTPChallenge } from "../../elements/stages/authenticator_totp/AuthenticatorTOTPStage";
-import { AuthenticatorStaticChallenge } from "../../elements/stages/authenticator_static/AuthenticatorStaticStage";
-import { AuthenticatorValidateStageChallenge } from "../../elements/stages/authenticator_validate/AuthenticatorValidateStage";
-import { WebAuthnAuthenticatorRegisterChallenge } from "../../elements/stages/authenticator_webauthn/WebAuthnAuthenticatorRegisterStage";
-import { CaptchaChallenge } from "../../elements/stages/captcha/CaptchaStage";
-import { COMMON_STYLES } from "../../common/styles";
-import { SpinnerSize } from "../../elements/Spinner";
-import { StageHost } from "../../elements/stages/base";
-import { Challenge, ChallengeTypeEnum, FlowsApi } from "../../api";
-import { DEFAULT_CONFIG } from "../../api/Config";
+import "./stages/authenticator_static/AuthenticatorStaticStage";
+import "./stages/authenticator_totp/AuthenticatorTOTPStage";
+import "./stages/authenticator_validate/AuthenticatorValidateStage";
+import "./stages/authenticator_webauthn/WebAuthnAuthenticatorRegisterStage";
+import "./stages/autosubmit/AutosubmitStage";
+import "./stages/captcha/CaptchaStage";
+import "./stages/consent/ConsentStage";
+import "./stages/email/EmailStage";
+import "./stages/identification/IdentificationStage";
+import "./stages/password/PasswordStage";
+import "./stages/prompt/PromptStage";
+import { ShellChallenge, RedirectChallenge } from "../api/Flows";
+import { IdentificationChallenge } from "./stages/identification/IdentificationStage";
+import { PasswordChallenge } from "./stages/password/PasswordStage";
+import { ConsentChallenge } from "./stages/consent/ConsentStage";
+import { EmailChallenge } from "./stages/email/EmailStage";
+import { AutosubmitChallenge } from "./stages/autosubmit/AutosubmitStage";
+import { PromptChallenge } from "./stages/prompt/PromptStage";
+import { AuthenticatorTOTPChallenge } from "./stages/authenticator_totp/AuthenticatorTOTPStage";
+import { AuthenticatorStaticChallenge } from "./stages/authenticator_static/AuthenticatorStaticStage";
+import { AuthenticatorValidateStageChallenge } from "./stages/authenticator_validate/AuthenticatorValidateStage";
+import { WebAuthnAuthenticatorRegisterChallenge } from "./stages/authenticator_webauthn/WebAuthnAuthenticatorRegisterStage";
+import { CaptchaChallenge } from "./stages/captcha/CaptchaStage";
+import { COMMON_STYLES } from "../common/styles";
+import { SpinnerSize } from "../elements/Spinner";
+import { StageHost } from "./stages/base";
+import { Challenge, ChallengeTypeEnum, FlowsApi } from "../api";
+import { DEFAULT_CONFIG } from "../api/Config";
 
 @customElement("ak-flow-executor")
 export class FlowExecutor extends LitElement implements StageHost {
