+++ web/src/flows/stages/identification/IdentificationStage.ts
@@ -4,6 +4,7 @@ import { Challenge } from "../../../api";
 import { COMMON_STYLES } from "../../../common/styles";
 import { BaseStage } from "../base";
 import "../form";
+import "../../../elements/utils/LoadingState";
 
 export interface IdentificationChallenge extends Challenge {
